Structure improvement
