+++ Sources/Results/PlacesResults.swift
@@ -1,0 +1,15 @@
+//
+//  PlacesResults.swift
+//  InstantSearchCore
+//
+//  Created by Vladislav Fitc on 28/08/2019.
+//  Copyright © 2019 Algolia. All rights reserved.
+//
+
+import Foundation
+
+public struct PlacesResults: Codable {
+  
+  let hits: [PlaceHit]
+  
+}
