+++ Sources/Searcher/Places/PlacesSearcher.swift
@@ -7,85 +7,6 @@
 //
 
 import Foundation
-import CoreLocation
-
-struct Geolocation: Codable {
-  
-  let latitude: Double
-  let longitude: Double
-  
-  enum CodingKeys: String, CodingKey {
-    case latitude = "lat"
-    case longitude = "lng"
-  }
-  
-}
-
-extension Geolocation: CustomDebugStringConvertible {
-  
-  var debugDescription: String {
-    return "{ lat: \(latitude), lon: \(longitude) }"
-  }
-  
-}
-
-extension Geolocation {
-  
-  init(_ coordinate: CLLocationCoordinate2D) {
-    self.init(latitude: coordinate.latitude, longitude: coordinate.longitude)
-  }
-  
-}
-
-extension CLLocationCoordinate2D {
-  
-  init(_ geolocation: Geolocation) {
-    self.init(latitude: geolocation.latitude, longitude: geolocation.longitude)
-  }
-  
-}
-
-struct PlaceHit: Codable {
-  
-  let id: String
-  let localeNames: [String]
-  let country: String
-  let county: [String]?
-  let administrative: [String]
-  let geolocation: Geolocation
-  
-  enum CodingKeys: String, CodingKey {
-    case id = "objectID"
-    case geolocation = "_geoloc"
-    case localeNames = "locale_names"
-    case country
-    case county
-    case administrative
-  }
-  
-}
-
-extension PlaceHit: CustomStringConvertible {
-  
-  var description: String {
-    return localeNames.first ?? ""
-  }
-  
-}
-
-extension PlaceHit: CustomDebugStringConvertible {
-  
-  var debugDescription: String {
-    return "{ id: \(id), locale names: \(localeNames), country: \(country), county: \(county ?? []), administrative: \(administrative), location: \(geolocation) }"
-  }
-  
-}
-
-public struct PlacesResults: Codable {
-  
-  let hits: [PlaceHit]
-  
-}
 
 public class PlacesSearcher: Searcher, SequencerDelegate, SearchResultObservable {
   
@@ -139,6 +60,7 @@ public class PlacesSearcher: Searcher, SequencerDelegate, SearchResultObservable
     self.onResults = .init()
     self.onError = .init()
     self.sequencer = .init()
+    placesQuery.language = "en"
     sequencer.delegate = self
     onResults.retainLastData = true
     isLoading.retainLastData = true
